+++ src/main/java/dev/frydae/beguild/user/RegisteredUser.java
@@ -12,7 +12,7 @@ import java.util.UUID;
 @Getter @Setter
 @RequiredArgsConstructor
 public class RegisteredUser {
-    @Expose private Long userId;
+    @Expose private Integer userId;
     @Expose private String name;
     @Expose private final UUID uuid;
     @Expose private Long lastLogin;
